Done all tasks

+++ python_variable_annotations/2-floor.py
@@ -3,5 +3,5 @@
 type-annotated function floor which takes a 
 float n as argument and returns the floor of the float.
 """
-def floor(n: float) -> int:
+def floor(n: float) -> float:
     return n
